dogs vs cats/requirements.txt
appdirs==1.4.4
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cryptography @ file:///croot/cryptography_1677533068310/work
filelock @ file:///croot/filelock_1672387128942/work
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
idna @ file:///croot/idna_1666125576474/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib @ file:///croot/joblib_1666298844297/work
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpmath==1.2.1
networkx @ file:///croot/networkx_1678964333703/work
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
packaging @ file:///croot/packaging_1678965309396/work
pandas==1.5.3
Pillow==9.4.0
pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///croot/pytz_1671697431263/work
requests @ file:///croot/requests_1682607517574/work
scikit-learn @ file:///croot/scikit-learn_1680198742267/work
scipy==1.10.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy @ file:///croot/sympy_1668202399572/work
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
torch==2.0.0
torchaudio==2.0.0
torchvision==0.15.0
triton==2.0.0
typing_extensions @ file:///croot/typing_extensions_1681939499988/work
urllib3 @ file:///croot/urllib3_1680254681959/work
